+++ src/superannotate/lib/app/serializers.py
@@ -166,6 +166,19 @@ class SettingsSerializer:
         return self.data
 
 
+class ItemSerializer(BaseSerializer):
+    def serialize(
+        self,
+        fields: List[str] = None,
+        by_alias: bool = False,
+        flat: bool = False,
+        exclude: Set[str] = None,
+    ):
+        data = super().serialize(fields, by_alias, flat, exclude)
+
+        return data
+
+
 class EntitySerializer:
     @classmethod
     def serialize(